Move `SpringInstrumentationContext` and `SpringUtExecutionInstrumentation` into separate package
